updated godthread to streamline child-killing. now it is part of the "run()" function of GodThread, so all child-killing is handled already by the abstract thread.
temporarily commented out "shitWentWrong" for two reasons:
1. its original purpose is to streamline child-killing, but that is now done as a "finally" statement in run()
2. its original purpose is to serve as a global error flag, but errors are not always global, and the godthreads should have more control over which errors it believes are fatal and which can be ignored. instead, error is now handled in the (soon-to-be-implemented) monitor threads, who will watch for specific errors and inform the godthread of specific violations

a temporary "delegateMonitor" function is added. this function currently assumes that all errors detected by the monitor threads are fatal. however, later on we may encounter errors that are not "fatal" and should be addressed differently. not sure if a generic "delegateMonitor" function will exist in the future.

renamed kill() to killActiveThreads() and added in killInnerThread(). killActiveThreads() is now private because the child-killing process is streamlined and it should therefore not be invoked anywhere else. killInnerThread() is protected so that it can be called in realRun() and its derivatives. killInnerThread() serves as a failsafe for terminating the autonomous code, in case something seriously goes wrong.